state_machine.c: use proper state variable to monitor the recorded higher priority MRA
MRP normative defines a specific retry counter to monitor the recorded
higher priority MRA. This patch introduces this counter and a related
variable to store the maximum value it can reach (this variable is
named MRP_MON_NR_MAX).

Unluckey the normative doesn't specify a recommended value for
MRP_MON_NR_MAX, so we set equal to Test Monitoring Count. This should
be a reasonable setting if the following disequation is valid:

    Default Test Interval
    --------------------- < Test Monitoring Count
    Short Test Interval

Signed-off-by: Rodolfo Giometti <giometti@enneenne.com>