state_machine.c: fix MIC forwarding rules
The MRP normative states that a node that behaves as MRC should not
forward interconnect frames between its ring ports if it has an
interconnection roles (MIM or MIC), but this is true for IN_TEST
frames having the same interconnection ring ID of the MIM or MIC
running on the node!

In fact, when a MIC, acting as MIM or MIC, receives an IN_TEST frame
with a different interconnection ring ID, then it should work as a
"simple" MIC (that is with no interconnection roles at all).

Signed-off-by: Rodolfo Giometti <giometti@enneenne.com>